Build: extract bin-links to a separate file for future package extraction.
