+++ lib/install/action/build.js
@@ -7,7 +7,7 @@ var packageId = require('../../utils/package-id.js')
 module.exports = function (staging, pkg, log, next) {
   log.silly('build', packageId(pkg))
   chain([
-    [build.linkStuff, pkg.package, pkg.path, npm.config.get('global'), true],
+    [build.linkStuff, pkg.package, pkg.path, npm.config.get('global')],
     [build.writeBuiltinConf, pkg.package, pkg.path]
   ], next)
 }
